change code according to comments

+++ server/controllers/masterControllers.js
@@ -6,7 +6,7 @@ class MasterController {
     postMaster = async (req, res) => {
         const errors = validationResult(req);
         if (!errors.isEmpty()) {
-            return res.status(400).json({ message: 'Uncorrect request', errors });
+            return res.status(400).json({ message: 'Incorrect request', errors });
         }
         try {
             const { name, rating, city } = req.body;
@@ -30,13 +30,13 @@ class MasterController {
     updateMaster = async (req, res) => {
         const errors = validationResult(req);
         if (!errors.isEmpty()) {
-            return res.status(400).json({ message: 'Uncorrect request', errors });
+            return res.status(400).json({ message: 'Incorrect request', errors });
         }
         try {
             const master = req.body;
             const { id } = req.params;
             if (!id) {
-                res.status(400).json({ message: 'ID не указан' });
+                res.status(400).json({ message: 'ID not found' });
             }
             await Master.findByIdAndUpdate(id, master, { new: true });
             const returnMaster = await Master.findById(id).populate({ path: 'city', select: 'city_name' });
@@ -47,18 +47,21 @@ class MasterController {
     };
 
     deleteMaster = async (req, res) => {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ message: 'Incorrect request', errors });
+        }
         try {
             const { id } = req.params;
-            if (!id) {
-                res.status(400).json({ message: 'ID не указан' });
-            }
+            /*if (!id) {
+                res.status(400).json({ message: 'ID not found' });
+            }*/
             const master = await Master.findByIdAndDelete(id);
 
             const allMasterOrders = master.order;
             for (let i = 0; i < allMasterOrders.length; i++) {
                 await Order.findByIdAndDelete(allMasterOrders[i]);
             }
-
             return res.json(master);
         } catch (e) {
             res.status(500).json(e);
